Extract classes, drop charset constant and use autoload
