+++ composer.json
@@ -18,10 +18,7 @@
     "autoload": {
         "psr-4": {
             "BazaarvoiceSeo\\": "src/"
-        },
-        "files": [
-            "src/bvseosdk.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
