+++ tests/Classes/BVUtilityTest.php
@@ -2,9 +2,6 @@
 
 namespace BazaarvoiceSeo;
 
-require_once 'src/BVUtility.php';
-require_once 'tests/config.php';
-
 /**
  * Test BVUtility class.
  */
