+++ tests/StoriesTest.php
@@ -1,8 +1,5 @@
 <?php
 
-require_once 'src/bvseosdk.php';
-require_once 'tests/config.php';
-
 /**
  * Test class to test Stories.
  */
